Feat: [BADA-340] 마감임박 페이지 무한 스크롤 적용 및 스와이퍼 속도 조절, 거래 페이지 리팩토링 #340
